refactor(ratewise): seo-metadata 設定模組化拆分
- 2959 行單檔拆為 seo-metadata/ 目錄（core + currency-landing + index barrel）
- import 路徑經 barrel 維持不變；GUIDE_LINK 常數改 export 供 currency 引用
- 同步守門測試／lastmod SSOT／markdown-mirrors prebuild 三類讀檔耦合

測試：typecheck、lint、test 2513 passed、build:ratewise 全綠零漂移

via [Happy](https://happy.engineering)
